Add EX Jaina
I'm also using this as a formatting test. The online version of Yomi refers to the EX characters as "EX Name," like when it says that I need to beat Yomibot Setsuki 25 times to unlock Yomibot EX Setsuki. Thus, having the EX as a subscript behind the name is incorrect, and if this test with EX Jaina's name formatting works, I'll apply it to EX Grave as well.