Add EX Midori
To do: Find a way to change the list of which cards are attacks, throws, blocks and dodges to include the Destiny cards for EX characters. As it is right now, the only way to tell that Midori has gained a new rank of card is to look at the ability list.