Fixed bugs and inconsistencies in Arg's combos
His good combos starting with normal attacks were bugged because they used "goodCombo" instead of "goodDamage," meaning they overwrote the actual combo cards with what was meant to be the damage (and displayed the rank of the normal as the damage to boot).

Also, his throw combos seemed to have been written with the assumption that he had 1 combo point throws, which is incorrect. I removed the extra attack from each of them.