Adding 1st Edition Jaina as a (WIP) variant
This required two changes to moves in addition to abilities, and I'm only able to do one of them.

The first is that Charged Shot was originally a Can't Combo move with 8.6 speed. Easy enough, though it does mean repeating the entire attack list to add that change...

The second is that her red normals originally did 3 block damage regardless of which rank you used. That takes a more substantial overhaul of how her normals are coded, and I don't know where to begin on that...